+++ examples/ccccl/runtime/ccccl_rt_internal.h
@@ -42,7 +42,7 @@ struct LObj {
     union {
         LObjAtom    atom;
         LObjPair    pair;
-        long        ival;
+        long long   ival;
         LObjClosure closure;
     } as;
 };
